Refactored suggested profiles to fully utilise tanstack query
