+++ apps/admin-x-activitypub/src/components/Search.tsx
@@ -13,8 +13,8 @@ import NiceModal from '@ebay/nice-modal-react';
 import ViewProfileModal from './modals/ViewProfileModal';
 
 import Separator from './global/Separator';
-import useSuggestedProfiles from '../hooks/useSuggestedProfiles';
-import {useSearchForUser} from '../hooks/useActivityPubQueries';
+
+import {useSearchForUser, useSuggestedProfiles} from '../hooks/useActivityPubQueries';
 
 interface SearchResultItem {
     actor: ActorProperties;
@@ -122,7 +122,9 @@ interface SearchProps {}
 
 const Search: React.FC<SearchProps> = ({}) => {
     // Initialise suggested profiles
-    const {suggested, isLoadingSuggested, updateSuggestedProfile} = useSuggestedProfiles(6);
+    const {suggestedProfilesQuery, updateSuggestedProfile} = useSuggestedProfiles('index', 6);
+    const {data: suggestedData, isLoading: isLoadingSuggested} = suggestedProfilesQuery;
+    const suggested = suggestedData || [];
 
     // Initialise search query
     const queryInputRef = useRef<HTMLInputElement>(null);
